Extract "Revision Diff" control from FormBrowse
